More ints

+++ ElgrosLib/Interfaces/IProductManager.cs
@@ -33,7 +33,7 @@ namespace ElgrosLib.Interfaces
         /// <param name="id"></param>
         /// <returns></returns>
         /// <exception cref="NotImplementedException"></exception>
-        public Task<Product> GetByIdAsync(long id);
+        public Task<Product> GetByIdAsync(int id);
 
         /// <summary>
         /// Updates a product
